gio: Use `Slice<u8>` in a few places instead of `Vec<u8>` to avoid copies
The returned values are already transfer full so copying them to a `Vec`
is unnecessary.
