+++ gio/Gir.toml
@@ -772,6 +772,16 @@ manual_traits = ["FileExtManual"]
     name = "new_tmp_dir_async"
     manual = true
     doc_trait_name = "FileExtManual"
+    [[object.function]]
+    name = "load_contents"
+    # Use `Slice<u8>`
+    manual = true
+    doc_trait_name = "FileExtManual"
+    [[object.function]]
+    name = "load_contents_async"
+    # Use `Slice<u8>`
+    manual = true
+    doc_trait_name = "FileExtManual"
 
 [[object]]
 name = "Gio.FileAttributeInfoList"
