+++ gio/src/auto/file.rs
@@ -903,117 +903,6 @@ pub trait FileExt: IsA<File> + sealed::Sealed + 'static {
         ))
     }
 
-    #[doc(alias = "g_file_load_contents")]
-    fn load_contents(
-        &self,
-        cancellable: Option<&impl IsA<Cancellable>>,
-    ) -> Result<(Vec<u8>, Option<glib::GString>), glib::Error> {
-        unsafe {
-            let mut contents = std::ptr::null_mut();
-            let mut length = std::mem::MaybeUninit::uninit();
-            let mut etag_out = std::ptr::null_mut();
-            let mut error = std::ptr::null_mut();
-            let is_ok = ffi::g_file_load_contents(
-                self.as_ref().to_glib_none().0,
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                &mut contents,
-                length.as_mut_ptr(),
-                &mut etag_out,
-                &mut error,
-            );
-            debug_assert_eq!(is_ok == glib::ffi::GFALSE, !error.is_null());
-            if error.is_null() {
-                Ok((
-                    FromGlibContainer::from_glib_full_num(contents, length.assume_init() as _),
-                    from_glib_full(etag_out),
-                ))
-            } else {
-                Err(from_glib_full(error))
-            }
-        }
-    }
-
-    #[doc(alias = "g_file_load_contents_async")]
-    fn load_contents_async<
-        P: FnOnce(Result<(Vec<u8>, Option<glib::GString>), glib::Error>) + 'static,
-    >(
-        &self,
-        cancellable: Option<&impl IsA<Cancellable>>,
-        callback: P,
-    ) {
-        let main_context = glib::MainContext::ref_thread_default();
-        let is_main_context_owner = main_context.is_owner();
-        let has_acquired_main_context = (!is_main_context_owner)
-            .then(|| main_context.acquire().ok())
-            .flatten();
-        assert!(
-            is_main_context_owner || has_acquired_main_context.is_some(),
-            "Async operations only allowed if the thread is owning the MainContext"
-        );
-
-        let user_data: Box_<glib::thread_guard::ThreadGuard<P>> =
-            Box_::new(glib::thread_guard::ThreadGuard::new(callback));
-        unsafe extern "C" fn load_contents_async_trampoline<
-            P: FnOnce(Result<(Vec<u8>, Option<glib::GString>), glib::Error>) + 'static,
-        >(
-            _source_object: *mut glib::gobject_ffi::GObject,
-            res: *mut crate::ffi::GAsyncResult,
-            user_data: glib::ffi::gpointer,
-        ) {
-            let mut error = std::ptr::null_mut();
-            let mut contents = std::ptr::null_mut();
-            let mut length = std::mem::MaybeUninit::uninit();
-            let mut etag_out = std::ptr::null_mut();
-            let _ = ffi::g_file_load_contents_finish(
-                _source_object as *mut _,
-                res,
-                &mut contents,
-                length.as_mut_ptr(),
-                &mut etag_out,
-                &mut error,
-            );
-            let result = if error.is_null() {
-                Ok((
-                    FromGlibContainer::from_glib_full_num(contents, length.assume_init() as _),
-                    from_glib_full(etag_out),
-                ))
-            } else {
-                Err(from_glib_full(error))
-            };
-            let callback: Box_<glib::thread_guard::ThreadGuard<P>> =
-                Box_::from_raw(user_data as *mut _);
-            let callback: P = callback.into_inner();
-            callback(result);
-        }
-        let callback = load_contents_async_trampoline::<P>;
-        unsafe {
-            ffi::g_file_load_contents_async(
-                self.as_ref().to_glib_none().0,
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                Some(callback),
-                Box_::into_raw(user_data) as *mut _,
-            );
-        }
-    }
-
-    fn load_contents_future(
-        &self,
-    ) -> Pin<
-        Box_<
-            dyn std::future::Future<Output = Result<(Vec<u8>, Option<glib::GString>), glib::Error>>
-                + 'static,
-        >,
-    > {
-        Box_::pin(crate::GioFuture::new(
-            self,
-            move |obj, cancellable, send| {
-                obj.load_contents_async(Some(cancellable), move |res| {
-                    send.resolve(res);
-                });
-            },
-        ))
-    }
-
     #[doc(alias = "g_file_make_directory")]
     fn make_directory(
         &self,
